add missing casts to avoid warnings
git-svn-id: https://svn.r-project.org/R/trunk@68743 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/main/connections.c
@@ -4515,11 +4515,11 @@ SEXP attribute_hidden do_readbin(SEXP call, SEXP op, SEXP args, SEXP env)
                         break;
 #if SIZEOF_LONG == 8
                     case sizeof(long):
-                        INTEGER(ans)[i] = u.l;
+                        INTEGER(ans)[i] = (int)u.l;
                         break;
 #elif SIZEOF_LONG_LONG == 8
                     case sizeof(_lli_t):
-                        INTEGER(ans)[i] = u.ll;
+                        INTEGER(ans)[i] = (int)u.ll;
                         break;
 #endif
                     default:
@@ -4548,7 +4548,7 @@ SEXP attribute_hidden do_readbin(SEXP call, SEXP op, SEXP args, SEXP env)
                         break;
 #if HAVE_LONG_DOUBLE && (SIZEOF_LONG_DOUBLE > SIZEOF_DOUBLE)
                     case sizeof(long double):
-                        REAL(ans)[i] = u.ld;
+                        REAL(ans)[i] = (double)u.ld;
                         break;
 #endif
                     default:
